+++ src/pool/disjoint/pool_disjoint_impl.hpp
